bootPLS-package

: Several implementations of non-parametric stable bootstrap-based techniques to determine the numbers of components for Partial Least Squares linear or generalized linear regression models as well as and sparse Partial Least Squares linear or generalized linear regression models. The package collects techniques that were published in a book chapter (Magnanensi et al. 2016, 'The Multiple Facets of Partial Least Squares and Related Methods', doi: 10.1007/978-3-319-40643-5_18 ) and two articles (Magnanensi et al. 2017, 'Statistics and Computing', doi: 10.1007/s11222-016-9651-4 ) and (Magnanensi et al. 2021, 'Frontiers in Applied Mathematics and Statistics', accepted.).

References

A new bootstrap-based stopping criterion in PLS component construction, J. Magnanensi, M. Maumy-Bertrand, N. Meyer and F. Bertrand (2016), in The Multiple Facets of Partial Least Squares and Related Methods, doi: 10.1007/978-3-319-40643-5_18

A new universal resample-stable bootstrap-based stopping criterion for PLS component construction, J. Magnanensi, F. Bertrand, M. Maumy-Bertrand and N. Meyer, (2017), Statistics and Computing, 27, 757–774. doi: 10.1007/s11222-016-9651-4

New developments in Sparse PLS regression, J. Magnanensi, M. Maumy-Bertrand, N. Meyer and F. Bertrand, (2021), Frontiers in Applied Mathematics and Statistics, accepted.

Author

Jérémy Magnanensi, Frédéric Bertrand
frederic.bertrand@utt.fr
https://fbertran.github.io/homepage/

Examples

set.seed(314)
library(bootPLS)
data(datasim)
head(datasim)

#>         Ygamma           X1          X2        X3          X4          X5
#> [1,] 1.7880705 -0.216611697 -0.17681134 0.5197978 -0.02799803  0.15292461
#> [2,] 3.9202532 -0.008908963  0.03080155 0.1287285  0.11680443  0.04339419
#> [3,] 0.8106974 -0.467074707  0.28725321 0.2308221  0.51380709  0.27666850
#> [4,] 1.5727608 -0.288192932 -0.11276107 0.5474216  0.02127908  0.20411778
#> [5,] 2.0274872 -0.255445454 -0.06399335 0.3905194  0.30555961 -0.03325430
#> [6,] 2.1619229 -0.179822610  0.12089428 0.1309757  0.16975015  0.18245005
#>               X6            X7
#> [1,] -0.08611683  0.1279287046
#> [2,]  0.10372045 -0.0001301158
#> [3,] -0.21949392 -0.4686856273
#> [4,] -0.12404658  0.0633289980
#> [5,]  0.10767176 -0.2427925800
#> [6,] -0.15040324 -0.1096520444