Bootstrap (Y,X) for the coefficients with number of components updated for each resampling.

Bootstrap (Y,X) for the coefficients with number of components updated for each resampling.

Usage

coefs.plsR.adapt.ncomp(
  dataset,
  i,
  R = 1000,
  ncpus = 1,
  parallel = "no",
  verbose = FALSE
)

Arguments

dataset

Dataset to use.

i

Vector of resampling.

R

Number of resamplings to find the number of components.

ncpus

integer: number of processes to be used in parallel operation: typically one would chose this to the number of available CPUs.

parallel

The type of parallel operation to be used (if any). If missing, the default is taken from the option "boot.parallel" (and if that is not set, "no").

verbose

Suppress information messages.

Value

Numeric vector: first value is the number of components, the remaining values are the coefficients the variables computed for that number of components.

References

A new bootstrap-based stopping criterion in PLS component construction, J. Magnanensi, M. Maumy-Bertrand, N. Meyer and F. Bertrand (2016), in The Multiple Facets of Partial Least Squares and Related Methods, doi:10.1007/978-3-319-40643-5_18

A new universal resample-stable bootstrap-based stopping criterion for PLS component construction, J. Magnanensi, F. Bertrand, M. Maumy-Bertrand and N. Meyer, (2017), Statistics and Computing, 27, 757–774. doi:10.1007/s11222-016-9651-4

New developments in Sparse PLS regression, J. Magnanensi, M. Maumy-Bertrand, N. Meyer and F. Bertrand, (2021), Frontiers in Applied Mathematics and Statistics, doi:10.3389/fams.2021.693126
.

Author

Jérémy Magnanensi, Frédéric Bertrand
frederic.bertrand@lecnam.net
https://fbertran.github.io/homepage/

Examples

set.seed(314)
ncol=5
xran=matrix(rnorm(30*ncol),30,ncol)
coefs.plsR.adapt.ncomp(xran,sample(1:30))
#> Loading required namespace: plsdof
#> [1]  0.0000000 -0.3144043  0.0000000  0.0000000  0.0000000  0.0000000
# \donttest{
coefs.plsR.adapt.ncomp(xran,sample(1:30),ncpus=2,parallel="multicore")
#> [1]  0.0000000 -0.3144043  0.0000000  0.0000000  0.0000000  0.0000000
# }