Permutation Bootstrap (Y,T) function for plsRglm

Permutation Bootstrap (Y,T) function for plsRglm

permcoefs.sgpls.CSim(
  dataRepYtt,
  ind,
  nt,
  modele,
  family = binomial,
  maxcoefvalues,
  ifbootfail
)

Arguments

dataRepYtt
Dataset with tt components to resample

ind
indices for resampling

nt
number of components to use

modele
type of modele to use, see plsRglm. Not used, please specify the family instead.

family
glm family to use, see plsRglm

maxcoefvalues
maximum values allowed for the estimates of the coefficients to discard those coming from singular bootstrap samples

ifbootfail
value to return if the estimation fails on a bootstrap sample

Value

Numeric vector of the components computed using a bootstrap resampling or ifbootfail value if the bootstrap computation fails.

References

A new bootstrap-based stopping criterion in PLS component construction, J. Magnanensi, M. Maumy-Bertrand, N. Meyer and F. Bertrand (2016), in The Multiple Facets of Partial Least Squares and Related Methods, doi: 10.1007/978-3-319-40643-5_18

A new universal resample-stable bootstrap-based stopping criterion for PLS component construction, J. Magnanensi, F. Bertrand, M. Maumy-Bertrand and N. Meyer, (2017), Statistics and Computing, 27, 757–774. doi: 10.1007/s11222-016-9651-4

New developments in Sparse PLS regression, J. Magnanensi, M. Maumy-Bertrand, N. Meyer and F. Bertrand, (2021), Frontiers in Applied Mathematics and Statistics, accepted.

Author

Jérémy Magnanensi, Frédéric Bertrand
frederic.bertrand@utt.fr
https://fbertran.github.io/homepage/

Examples

set.seed(4619)
xran=cbind(rbinom(30,1,.2),matrix(rnorm(150),30,5))
permcoefs.sgpls.CSim(xran, ind=sample(1:nrow(xran)), maxcoefvalues=1e5, 
ifbootfail=rep(NA,3))

#>        Tb1        Tb2        Tb3        Tb4        Tb5 
#> -0.2608434  0.7512146 -4.2041446 -2.2336737 -0.2452805